Rename `Syntax` -> `Parser`; `ByteFormat` -> `ByteParser`; and move `Literal` to crate root.
