+++ README.md
@@ -11,17 +11,16 @@ The term "parser composition" emphasizes how sophisticated parsers can be define
 ```rust
 use incpa::BaseParserError;
 use incpa::primitive::remaining;
-use incpa::syntax::Parser;
+use incpa::Parser;
 
 fn main() -> Result<(), BaseParserError> {
-    let syntax = define_syntax();
-    let input = "Hello World!";
-    let output = syntax.parse_all(input)?;
+    let parser = define_my_parser();
+    let output = parser.parse_all("Hello World!")?;
     assert_eq!(output, ("Hello", " World!".to_string()));
     Ok(())
 }
 
-fn define_syntax() -> impl Parser<str, Output=(&'static str, String), Error=BaseParserError> {
+fn define_my_parser() -> impl Parser<str, Output=(&'static str, String), Error=BaseParserError> {
     "Hello".then(remaining())
 }
 ```
@@ -40,7 +39,7 @@ This crate is in the version 0.0.x phase of early proof of concept with unstable
 
 ### 0.1.0 Feature Goals
 
-- [ ] A basic suite of general composition abstractions such as [Parser::then] and `Parser::or` with backtracking support.
+- [ ] A basic suite of general composition abstractions such as [Parser::then] and [Parser::or] with backtracking support.
 - [ ] Support for both string parsers and slice parsers (including byte slices)
 - [ ] Efficient streaming string parsing from byte-oriented I/O sources using UTF8 decoding
 - [ ] Common generic primitive parsers, such as end-of-input, constants, and literals
