+++ src/combinators/mapoutput.rs
