+++ src/combinators/then/tests.rs
@@ -4,9 +4,8 @@ use test_case::test_case;
 
 use crate::parsing::Buffer;
 use crate::primitive::remaining;
-use crate::syntax::Literal;
 use crate::testutils::test_buffer_windows_output_no_res;
-use crate::Parser;
+use crate::{Literal, Parser};
 
 #[test_case("hello world!")]
 #[test_case(b"hello world!".as_slice())]
