+++ src/lib.rs
@@ -1,11 +1,16 @@
 #![doc = include_str!("../README.md")]
 #![deny(missing_docs, unsafe_code)]
 
+mod byteparser;
+pub mod combinators;
 mod error;
+mod literal;
+mod parser;
 pub mod parsing;
 pub mod primitive;
-pub mod syntax;
 pub mod testutils;
 
+pub use self::byteparser::ByteParser;
 pub use self::error::BaseParserError;
-pub use self::syntax::Parser;
+pub use self::literal::Literal;
+pub use self::parser::Parser;
