+++ src/parser.rs
@@ -1,18 +1,4 @@
-//! [Parser] and impls for composing syntaxes to specify a parser's behavior
-mod byteformat;
-mod literal;
-mod maperror;
-mod mapoutput;
-mod or;
-mod then;
-
-pub use self::byteformat::ByteFormat;
-pub use self::literal::{Literal, LiteralParser};
-pub use self::maperror::MapError;
-pub use self::mapoutput::MapOutput;
-pub use self::or::Or;
-pub use self::then::Then;
-
+use crate::combinators::{MapError, MapOutput, Or, Then};
 use crate::parsing::{Buffer, ParserState};
 use crate::BaseParserError;
