+++ src/parsing/backtrack.rs
@@ -4,7 +4,7 @@ use crate::parsing::{Buffer, ParserState, Update};
 
 /// Try to parse `P`, but hold all input until a successful parse
 ///
-/// This ensures if `P` fails with an error, no input will have been consumed. Typically this is used internally for conditional parsing, such as with [syntax::Or](crate::syntax::Or).
+/// This ensures if `P` fails with an error, no input will have been consumed. Typically this is used internally for conditional parsing, such as with [combinators::Or](crate::combinators::Or).
 #[derive(Copy, Clone, Debug, new)]
 pub struct Backtrack<P> {
     inner: P,
